Moved test files to uport-defaults module
